fix: Cooking over Time
Coroutine sitzt jetzt auf den jeweiligen Objekten. 
Muss noch getestet werden.
