+++ Assets/SO_Dishes/Burger.asset
@@ -1,0 +1,17 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 725cd17fdbdba0c43a17cf203f5bb4a7, type: 3}
+  m_Name: Burger
+  m_EditorClassIdentifier: 
+  name: Burger
+  points: 100
+  dishArt: {fileID: 0}
