+++ Assets/SO_Dishes/Egg.asset
@@ -10,9 +10,8 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 725cd17fdbdba0c43a17cf203f5bb4a7, type: 3}
-  m_Name: CheeseBurger
+  m_Name: Egg
   m_EditorClassIdentifier: 
-  name: CheeseBurger
-  points: 10
-  dishArt: {fileID: 21300000, guid: 16e82af8ace150b4e969a5aed8fdac15, type: 3}
-  ingredients: 00000000030000000200000000000000
+  name: Egg
+  points: 25
+  dishArt: {fileID: 0}
