+++ Assets/SO_Dishes/Egg.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f101a398527cc47458f18c725733adba
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
