+++ Assets/SO_Dishes/Steak.asset
@@ -10,9 +10,8 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 725cd17fdbdba0c43a17cf203f5bb4a7, type: 3}
-  m_Name: Fries
+  m_Name: Steak
   m_EditorClassIdentifier: 
-  name: Fries
-  points: 10
-  dishArt: {fileID: 21300000, guid: 89d96de5938b4fb49b0533aacd9be16b, type: 3}
-  ingredients: 04000000
+  name: Steak
+  points: 50
+  dishArt: {fileID: 0}
