+++ Assets/SO_Dishes/Steak.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: da01fcf26a4f5df419c6293dcaa82c9d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
