+++ Assets/Scripts/ScriptableObject/DishScriptableObject.cs
@@ -10,14 +10,4 @@ public class DishScriptableObject : ScriptableObject
     public new string name;
     public int points;
     public Sprite dishArt;
-    public List<avalaibleIngredients> ingredients;
-    
-    public enum avalaibleIngredients
-    {
-        bred,
-        salat,
-        ham,
-        cheese,
-        fries
-    }
 }
