+++ Assets/Scripts/VR-Player/FryFoodOverTime.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5c78a44de666dd34f833e06d245d99de
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
